backup/requirements.txt
python-swiftclient==3.8.1
certifi==2019.11.28
chardet==3.0.4
idna==2.8
pip==19.0.3
requests==2.22.0
setuptools==40.8.0
six==1.14.0
urllib3==1.25.8
